Added main for a lot of function

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: mchauvin <mchauvin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/04 14:18:05 by mchauvin          #+#    #+#             */
-/*   Updated: 2025/11/13 15:36:55 by mchauvin         ###   ########.fr       */
+/*   Updated: 2025/11/17 10:33:47 by mchauvin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,8 +33,8 @@ void				*ft_memset(void *s, int v, size_t n);
 void				ft_bzero(void *s, size_t n);
 void				*ft_memcpy(void *dest, const void *src, size_t n);
 void				*ft_memmove(void *dest, const void *src, size_t n);
-unsigned int		ft_strlcpy(char *dest, const char *src, unsigned int size);
-unsigned int		ft_strlcat(char *dest, const char *src, unsigned int size);
+size_t				ft_strlcpy(char *dest, const char *src, size_t size);
+size_t				ft_strlcat(char *dest, const char *src, size_t size);
 int					ft_toupper(int c);
 int					ft_tolower(int c);
 char				*ft_strchr(const char *str, int to_find);
